+++ tests/TestCase/Driver/MysqlTest.php
@@ -15,8 +15,10 @@ declare(strict_types=1);
  */
 namespace DatabaseBackup\Test\TestCase\Driver;
 
+use DatabaseBackup\Driver\Driver;
 use DatabaseBackup\Driver\Mysql;
 use DatabaseBackup\TestSuite\DriverTestCase;
+use Tools\Filesystem;
 
 /**
  * MysqlTest class
@@ -24,16 +26,23 @@ use DatabaseBackup\TestSuite\DriverTestCase;
 class MysqlTest extends DriverTestCase
 {
     /**
+     * @var \DatabaseBackup\Driver\Mysql&\PHPUnit\Framework\MockObject\MockObject
+     */
+    protected Driver $Driver;
+
+    /**
      * Called before every test method
      * @return void
      */
     public function setUp(): void
     {
-        parent::setUp();
-
-        if (!$this->Driver instanceof Mysql) {
-            $this->markTestSkipped('Skipping tests for Mysql, current driver is ' . $this->Driver->getDriverName());
+        if ($this->getConnection()->config()['scheme'] !== 'mysql') {
+            $this->markTestSkipped('Skipping tests for mysql, current driver is `' . $this->getConnection()->config()['scheme'] . '`');
         }
+
+        $this->Driver ??= $this->createPartialMock(Mysql::class, ['getAuthFilePath', 'writeAuthFile']);
+
+        parent::setUp();
     }
 
     /**
@@ -42,9 +51,14 @@ class MysqlTest extends DriverTestCase
      */
     public function testAfterExport(): void
     {
-        $Driver = $this->createPartialMock(Mysql::class, ['deleteAuthFile']);
-        $Driver->expects($this->once())->method('deleteAuthFile');
-        $Driver->afterExport();
+        $expectedAuthFile = Filesystem::createTmpFile();
+        $this->assertFileExists($expectedAuthFile);
+
+        $Driver = $this->createPartialMock(Mysql::class, ['getAuthFilePath']);
+        $Driver->method('getAuthFilePath')->willReturn($expectedAuthFile);
+        $Driver->getEventManager()->on($Driver);
+        $Driver->dispatchEvent('Backup.afterExport');
+        $this->assertFileDoesNotExist($expectedAuthFile);
     }
 
     /**
@@ -53,9 +67,14 @@ class MysqlTest extends DriverTestCase
      */
     public function testAfterImport(): void
     {
-        $Driver = $this->createPartialMock(Mysql::class, ['deleteAuthFile']);
-        $Driver->expects($this->once())->method('deleteAuthFile');
-        $Driver->afterImport();
+        $expectedAuthFile = Filesystem::createTmpFile();
+        $this->assertFileExists($expectedAuthFile);
+
+        $Driver = $this->createPartialMock(Mysql::class, ['getAuthFilePath']);
+        $Driver->getEventManager()->on($Driver);
+        $Driver->method('getAuthFilePath')->willReturn($expectedAuthFile);
+        $Driver->dispatchEvent('Backup.afterImport');
+        $this->assertFileDoesNotExist($expectedAuthFile);
     }
 
     /**
@@ -64,16 +83,17 @@ class MysqlTest extends DriverTestCase
      */
     public function testBeforeExport(): void
     {
-        $this->assertTrue($this->Driver->beforeExport());
+        $expectedContent = '[mysqldump]' . PHP_EOL .
+            'user={{USER}}' . PHP_EOL .
+            'password="{{PASSWORD}}"' . PHP_EOL .
+            'host={{HOST}}';
 
-        $expected = '[mysqldump]' . PHP_EOL .
-            'user=' . $this->Driver->getConfig('username') . PHP_EOL .
-            'password="' . $this->Driver->getConfig('password') . '"' . PHP_EOL .
-            'host=' . $this->Driver->getConfig('host');
-        $auth = $this->invokeMethod($this->Driver, 'getAuthFile');
-        $this->assertStringEqualsFile($auth, $expected);
+        $this->Driver->expects($this->once())
+            ->method('writeAuthFile')
+            ->with($this->equalTo($expectedContent))
+            ->willReturn(true);
 
-        unlink($auth);
+        $this->assertTrue($this->Driver->dispatchEvent('Backup.beforeExport')->getResult());
     }
 
     /**
@@ -82,31 +102,39 @@ class MysqlTest extends DriverTestCase
      */
     public function testBeforeImport(): void
     {
-        $this->assertTrue($this->Driver->beforeImport());
+        $expectedContent = '[client]' . PHP_EOL .
+            'user={{USER}}' . PHP_EOL .
+            'password="{{PASSWORD}}"' . PHP_EOL .
+            'host={{HOST}}';
 
-        $expected = '[client]' . PHP_EOL .
-            'user=' . $this->Driver->getConfig('username') . PHP_EOL .
-            'password="' . $this->Driver->getConfig('password') . '"' . PHP_EOL .
-            'host=' . $this->Driver->getConfig('host');
-        $auth = $this->invokeMethod($this->Driver, 'getAuthFile');
-        $this->assertStringEqualsFile($auth, $expected);
+        $this->Driver->expects($this->once())
+            ->method('writeAuthFile')
+            ->with($this->equalTo($expectedContent))
+            ->willReturn(true);
 
-        unlink($auth);
+        $this->assertTrue($this->Driver->dispatchEvent('Backup.beforeImport')->getResult());
     }
 
     /**
      * @test
-     * @uses \DatabaseBackup\Driver\Mysql::deleteAuthFile()
+     * @uses \DatabaseBackup\Driver\Mysql::writeAuthFile()
      */
-    public function testDeleteAuthFile(): void
+    public function testWriteAuthFile(): void
     {
-        $this->assertFalse($this->invokeMethod($this->Driver, 'deleteAuthFile'));
+        $expectedAuthFile = TMP . 'auth' . uniqid();
+        $this->assertFileDoesNotExist($expectedAuthFile);
 
-        $auth = tempnam(sys_get_temp_dir(), 'auth') ?: '';
-        $Driver = $this->createPartialMock(Mysql::class, ['getAuthFile']);
-        $Driver->method('getAuthFile')->willReturn($auth);
-        $this->assertFileExists($auth);
-        $this->assertTrue($this->invokeMethod($Driver, 'deleteAuthFile'));
-        $this->assertFileDoesNotExist($auth);
+        $Driver = $this->createPartialMock(Mysql::class, ['getAuthFilePath']);
+        $Driver->method('getAuthFilePath')->willReturn($expectedAuthFile);
+        $Driver->getEventManager()->on($Driver);
+
+        //Dispatches an event that calls and returns `writeAuthFile()`
+        $this->assertTrue($Driver->dispatchEvent('Backup.beforeExport')->getResult());
+        $this->assertFileExists($expectedAuthFile);
+        $config = $Driver->getConnection()->config();
+        $this->assertSame('[mysqldump]' . PHP_EOL .
+            'user=' . $config['username'] . PHP_EOL .
+            'password="' . ($config['password'] ?? '') . '"' . PHP_EOL .
+            'host=' . $config['host'], file_get_contents($expectedAuthFile));
     }
 }