+++ tests/TestCase/Driver/PostgresTest.php
@@ -1,5 +1,4 @@
 <?php
-/** @noinspection PhpDocMissingThrowsInspection, PhpUnhandledExceptionInspection */
 declare(strict_types=1);
 
 /**
@@ -15,7 +14,6 @@ declare(strict_types=1);
  */
 namespace DatabaseBackup\Test\TestCase\Driver;
 
-use DatabaseBackup\Driver\Postgres;
 use DatabaseBackup\TestSuite\DriverTestCase;
 
 /**
@@ -31,8 +29,8 @@ class PostgresTest extends DriverTestCase
     {
         parent::setUp();
 
-        if (!$this->Driver instanceof Postgres) {
-            $this->markTestSkipped('Skipping tests for Postgres, current driver is ' . $this->Driver->getDriverName());
+        if ($this->getConnection()->config()['scheme'] !== 'postgres') {
+            $this->markTestSkipped('Skipping tests for postgres, current driver is `' . $this->getConnection()->config()['scheme'] . '`');
         }
     }
 }